Employed NetcdfDatasetCache in both Java (DataReader) and Jython (nj22dataset.py) code
git-svn-id: https://ncwms.svn.sourceforge.net/svnroot/ncwms/trunk@67 f020452a-5e1f-0410-8fa4-ae2b6fb27d5e

+++ src/java/uk/ac/rdg/resc/ncwms/datareader/DatasetCache.java
@@ -160,10 +160,10 @@ public class DatasetCache
                     {
                         tPos = findDimensionPos(dims, coordSys.getTimeAxis().getDimension(0));
                         Date[] tVals = coordSys.getTimeDates();
-                        float[] sse = new float[tVals.length]; // Seconds since the epoch
+                        double[] sse = new double[tVals.length]; // Seconds since the epoch
                         for (int i = 0; i < tVals.length; i++)
                         {
-                            sse[i] = tVals[i].getTime() / 1000.0f;
+                            sse[i] = tVals[i].getTime() / 1000.0;
                         }
                         vm.setTvalues(sse);
                     }
